requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
frozenlist==1.6.0
graphviz==0.20.3
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
marshmallow==3.26.1
mcp==1.8.0
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.77.0
openai-agents==0.0.14
orjson==3.10.18
packaging==24.2
pbs-installer==2025.4.9
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.3
poetry-core==2.1.3
propcache==0.3.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyproject_hooks==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.1
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.5.1.12
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.30.0
websockets==15.0.1
wheel==0.45.1
xattr==1.1.4
yarl==1.20.0
zstandard==0.23.0
